chore: move types.ts to common.ts
