+++ frontend/src/plugins/issue/logic/initialize/standard.ts
@@ -1,9 +1,8 @@
 import { TemplateType } from "@/plugins";
 import { IssueCreate, MigrationType } from "@/types";
-import { findDatabaseListByQuery } from "../common";
+import { findDatabaseListByQuery, BuildNewIssueContext } from "../common";
 import { ESTABLISH_BASELINE_SQL, VALIDATE_ONLY_SQL } from "./constants";
 import { IssueCreateHelper } from "./helper";
-import { BuildNewIssueContext } from "../types";
 
 export const buildNewStandardIssue = async (
   context: BuildNewIssueContext
